NormalizeProcessorConfig -> NormalizationConfig + move it
