--- relay-event-normalization/src/normalize/processor.rs
@@ -1,143 +1,0 @@
-//! Event normalization processor.
-//!
-//! This processor is work in progress. The intention is to have a single
-//! processor to deal with all event normalization. Currently, the normalization
-//! logic is split across several processing steps running at different times
-//! and under different conditions, like light normalization and store
-//! processing. Having a single processor will make things simpler.
-
-use std::ops::Range;
-
-use chrono::{DateTime, Utc};
-use relay_common::time::UnixTimestamp;
-use relay_event_schema::protocol::IpAddr;
-
-use crate::{
-    BreakdownsConfig, DynamicMeasurementsConfig, GeoIpLookup, PerformanceScoreConfig,
-    RawUserAgentInfo, SpanDescriptionRule, TransactionNameConfig,
-};
-
-/// Configuration for [`NormalizeProcessor`].
-#[derive(Clone, Debug)]
-pub struct NormalizeProcessorConfig<'a> {
-    /// The IP address of the SDK that sent the event.
-    ///
-    /// When `{{auto}}` is specified and there is no other IP address in the payload, such as in the
-    /// `request` context, this IP address gets added to the `user` context.
-    pub client_ip: Option<&'a IpAddr>,
-
-    /// The user-agent and client hints obtained from the submission request headers.
-    ///
-    /// Client hints are the preferred way to infer device, operating system, and browser
-    /// information should the event payload contain no such data. If no client hints are present,
-    /// normalization falls back to the user agent.
-    pub user_agent: RawUserAgentInfo<&'a str>,
-
-    /// The time at which the event was received in this Relay.
-    ///
-    /// This timestamp is persisted into the event payload.
-    pub received_at: Option<DateTime<Utc>>,
-
-    /// The maximum amount of seconds an event can be dated in the past.
-    ///
-    /// If the event's timestamp is older, the received timestamp is assumed.
-    pub max_secs_in_past: Option<i64>,
-
-    /// The maximum amount of seconds an event can be predated into the future.
-    ///
-    /// If the event's timestamp lies further into the future, the received timestamp is assumed.
-    pub max_secs_in_future: Option<i64>,
-
-    /// Timestamp range in which a transaction must end.
-    ///
-    /// Transactions that finish outside of this range are considered invalid.
-    /// This check is skipped if no range is provided.
-    pub transaction_range: Option<Range<UnixTimestamp>>,
-
-    /// The maximum length for names of custom measurements.
-    ///
-    /// Measurements with longer names are removed from the transaction event and replaced with a
-    /// metadata entry.
-    pub max_name_and_unit_len: Option<usize>,
-
-    /// Configuration for measurement normalization in transaction events.
-    ///
-    /// Has an optional [`crate::MeasurementsConfig`] from both the project and the global level.
-    /// If at least one is provided, then normalization will truncate custom measurements
-    /// and add units of known built-in measurements.
-    pub measurements: Option<DynamicMeasurementsConfig<'a>>,
-
-    /// Emit breakdowns based on given configuration.
-    pub breakdowns_config: Option<&'a BreakdownsConfig>,
-
-    /// When `Some(true)`, context information is extracted from the user agent.
-    pub normalize_user_agent: Option<bool>,
-
-    /// Configuration to apply to transaction names, especially around sanitizing.
-    pub transaction_name_config: TransactionNameConfig<'a>,
-
-    /// When `Some(true)`, it is assumed that the event has been normalized before.
-    ///
-    /// This disables certain normalizations, especially all that are not idempotent. The
-    /// renormalize mode is intended for the use in the processing pipeline, so an event modified
-    /// during ingestion can be validated against the schema and large data can be trimmed. However,
-    /// advanced normalizations such as inferring contexts or clock drift correction are disabled.
-    ///
-    /// `None` equals to `false`.
-    pub is_renormalize: bool,
-
-    /// When `true`, infers the device class from CPU and model.
-    pub device_class_synthesis_config: bool,
-
-    /// When `true`, extracts tags from event and spans and materializes them into `span.data`.
-    pub enrich_spans: bool,
-
-    /// When `true`, computes and materializes attributes in spans based on the given configuration.
-    pub light_normalize_spans: bool,
-
-    /// The maximum allowed size of tag values in bytes. Longer values will be cropped.
-    pub max_tag_value_length: usize, // TODO: move span related fields into separate config.
-
-    /// Configuration for replacing identifiers in the span description with placeholders.
-    ///
-    /// This is similar to `transaction_name_config`, but applies to span descriptions.
-    pub span_description_rules: Option<&'a Vec<SpanDescriptionRule>>,
-
-    /// Configuration for generating performance score measurements for web vitals
-    pub performance_score: Option<&'a PerformanceScoreConfig>,
-
-    /// An initialized GeoIP lookup.
-    pub geoip_lookup: Option<&'a GeoIpLookup>,
-
-    /// When `Some(true)`, individual parts of the event payload is trimmed to a maximum size.
-    ///
-    /// See the event schema for size declarations.
-    pub enable_trimming: bool,
-}
-
-impl<'a> Default for NormalizeProcessorConfig<'a> {
-    fn default() -> Self {
-        Self {
-            client_ip: Default::default(),
-            user_agent: Default::default(),
-            received_at: Default::default(),
-            max_secs_in_past: Default::default(),
-            max_secs_in_future: Default::default(),
-            transaction_range: Default::default(),
-            max_name_and_unit_len: Default::default(),
-            breakdowns_config: Default::default(),
-            normalize_user_agent: Default::default(),
-            transaction_name_config: Default::default(),
-            is_renormalize: Default::default(),
-            device_class_synthesis_config: Default::default(),
-            enrich_spans: Default::default(),
-            light_normalize_spans: Default::default(),
-            max_tag_value_length: usize::MAX,
-            span_description_rules: Default::default(),
-            performance_score: Default::default(),
-            geoip_lookup: Default::default(),
-            enable_trimming: false,
-            measurements: None,
-        }
-    }
-}
